Fix upvote toggle issue
Upvote icon was not toggling properly on pages where the same bug could be displayed in two places. (eg. the profile page, where the user could have a bug showing in the 'my issues' tab, and the 'bugs I've upvoted' tag. Fixed using a custom attribute instead of bug id